chore: seperate object result and object result matching
