+++ perception_eval/test/visualization/test_eda_tool.py
@@ -29,7 +29,7 @@ from perception_eval.common.shape import Shape
 from perception_eval.common.shape import ShapeType
 from perception_eval.evaluation.matching.object_matching import MatchingMode
 from perception_eval.evaluation.result.object_result import DynamicObjectWithPerceptionResult
-from perception_eval.evaluation.result.object_result import get_object_results
+from perception_eval.evaluation.result.object_result_matching import get_object_results
 from perception_eval.util.debug import get_objects_with_difference
 from perception_eval.visualization.eda_tool import EDAManager
 from perception_eval.visualization.eda_tool import EDAVisualizer
